refactor(metrics): split the accelerated-mass concept out of divot.py
Two changed-file budgets were over after the #8659 work: divot.py at
1201 lines against 1200, and divot_metrics at 115 against 100. Both are
symptoms of the same thing -- an interval-valued mass correction, with
its own provenance and its own lateral-spread model, grew inside a
module about divot geometry.

Moved to a new metrics/accelerated_mass.py: AcceleratedSandMass,
lateral_spread_factor, F1_ENTRAINMENT_FACTOR_BOUNDS and the two reason
strings. That is a concept boundary, not a line-count trick -- the
accelerated mass is a consistency correction between two uncalibrated
models, while divot.py measures the geometry a sole actually swept.
divot.py re-exports all five names, so every existing import keeps
working.

divot_metrics then lost its input guards to _require_divot_inputs and
its two trapezoid integrals to _section_integrals. The integrals belong
together because lateral_spread_factor needs the squared-depth integral
on exactly the window the section area was taken over; separating them
would invite the two drifting onto different windows.

divot.py 1201 -> 1075, divot_metrics 115 -> 92, both inside budget with
room. 319 tests across metrics/ and ball/ pass unchanged; ruff, format
and the architecture budget gate are clean.
